requirements.txt
autopep8==1.6.0
celery==5.2.7
channels==3.0.5
channels-redis==4.0.0
Django==4.1.2
django-celery-results==2.4.0
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
flower==1.2.0
jsonschema==2.6.0
Markdown==3.4.1
mqttasgi==1.1.2
Pillow==9.2.0
redis==4.3.4
requests==2.28.1
psycopg2-binary